+++ Tests/Fixtures/Model/ModelCustomTranslatable.php
@@ -1,0 +1,7 @@
+<?php
+
+namespace Sonata\TranslationBundle\Tests\Fixtures\Model;
+
+class ModelCustomTranslatable
+{
+}
