+++ Tests/Fixtures/Model/ModelPersonalTranslatable.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace Sonata\TranslationBundle\Tests\Fixtures\Model;
+
+use Sonata\TranslationBundle\Model\Gedmo\AbstractPersonalTranslatable;
+use Sonata\TranslationBundle\Model\Gedmo\TranslatableInterface;
+
+class ModelPersonalTranslatable extends AbstractPersonalTranslatable implements TranslatableInterface
+{
+}
