+++ Tests/Fixtures/Model/ModelPersonalTranslation.php
@@ -1,0 +1,9 @@
+<?php
+
+namespace Sonata\TranslationBundle\Tests\Fixtures\Model;
+
+use Sonata\TranslationBundle\Model\Gedmo\AbstractPersonalTranslation;
+
+class ModelPersonalTranslation extends AbstractPersonalTranslation
+{
+}
