+++ Tests/Fixtures/Model/ModelTranslatable.php
@@ -1,0 +1,11 @@
+<?php
+
+namespace Sonata\TranslationBundle\Tests\Fixtures\Model;
+
+use Sonata\TranslationBundle\Model\Gedmo\AbstractTranslatable;
+use Sonata\TranslationBundle\Model\Gedmo\TranslatableInterface;
+
+class ModelTranslatable extends AbstractTranslatable implements TranslatableInterface
+{
+}
+
