+++ Tests/Fixtures/Model/ModelUsingTraitTranslatable.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace Sonata\TranslationBundle\Tests\Fixtures\Model;
+
+use Sonata\TranslationBundle\Traits\TranslatableTrait;
+
+class ModelUsingTraitTranslatable
+{
+    use TranslatableTrait;
+}
