+++ Tests/Fixtures/Traits/ModelPersonalTranslatable.php
@@ -1,0 +1,22 @@
+<?php
+
+namespace Sonata\TranslationBundle\Tests\Fixtures\Traits;
+
+use Doctrine\Common\Collections\ArrayCollection;
+use Sonata\TranslationBundle\Model\Gedmo\TranslatableInterface;
+use Sonata\TranslationBundle\Traits\Gedmo\PersonalTranslatableTrait;
+
+class ModelPersonalTranslatable implements TranslatableInterface
+{
+    use PersonalTranslatableTrait;
+
+    /**
+     * @var ArrayCollection
+     */
+    protected $translations;
+
+    public function __construct()
+    {
+        $this->translations = new ArrayCollection();
+    }
+}
