+++ Tests/Fixtures/Traits/ModelPersonalTranslation.php
@@ -1,0 +1,9 @@
+<?php
+
+namespace Sonata\TranslationBundle\Tests\Fixtures\Traits;
+
+use Sonata\TranslationBundle\Model\Gedmo\AbstractPersonalTranslation;
+
+class ModelPersonalTranslation extends AbstractPersonalTranslation
+{
+}
