+++ Tests/Fixtures/Traits/ModelTranslatable.php
@@ -1,0 +1,11 @@
+<?php
+
+namespace Sonata\TranslationBundle\Tests\Fixtures\Traits;
+
+use Sonata\TranslationBundle\Model\Gedmo\TranslatableInterface;
+use Sonata\TranslationBundle\Traits\TranslatableTrait;
+
+class ModelTranslatable implements TranslatableInterface
+{
+    use TranslatableTrait;
+}
